refactor(studio): decompose useGsapScriptCommits into focused mutation hooks
